requirements.txt
numpy==1.22.4
pandas==1.4.2
st_annotated_text==3.0.0
streamlit==1.9.2
